Chore: Refactoring datasets.py step 1 - stimuli in own file
